make DumpPublicKeys write 32-bit values as unsigned
Write the 32-bit chunks of N and RR as unsigned, so we get:

  {64,0xc926ad21,{1795090719,2141396315,950055447,2581568430,4268923165,...

instead of:

  {64,0xc926ad21,{1795090719,2141396315,950055447,-1713398866,-26044131,...

+++ libmincrypt/tools/DumpPublicKey.java
@@ -77,7 +77,7 @@ class DumpPublicKey {
         // Write out modulus as little endian array of integers.
         result.append(",{");
         for (int i = 0; i < nwords; ++i) {
-            int n = N.mod(B).intValue();
+            long n = N.mod(B).longValue();
             result.append(n);
 
             if (i != nwords - 1) {
@@ -91,7 +91,7 @@ class DumpPublicKey {
         // Write R^2 as little endian array of integers.
         result.append(",{");
         for (int i = 0; i < nwords; ++i) {
-            int rr = RR.mod(B).intValue();
+            long rr = RR.mod(B).longValue();
             result.append(rr);
 
             if (i != nwords - 1) {
